backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
Pillow>=10.2.0
numpy>=1.26.4
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
ffmpeg-python==0.2.0
minio==7.2.15
psycopg2-binary==2.9.9
asyncpg>=0.29.0
aiosqlite>=0.20.0
alembic==1.13.1
